HTTP/3.0/requirements.txt
aioquic==0.9.22
jwt==1.3.1
opentelemetry-sdk==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-instrumentation==0.40b0
